requirements.txt
absl-py==0.11.0
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
atpublic==2.1.3
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
confuse==1.4.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dictdiffer==0.8.1
distro==1.5.0
dpath==2.0.1
dulwich==0.20.19
dvc==1.11.16
entrypoints==0.3
flatbuffers==1.12
flatten-dict==0.3.0
flufl.lock==3.2
ftfy==5.9
funcy==1.15
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.13
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grandalf==0.6
grpcio==1.32.0
h5py==2.10.0
htmlmin==0.1.12
idna==2.10
ImageHash==4.2.0
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonpath-ng==1.5.2
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.35.0
mailchecker==4.0.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
missingno==0.4.2
mistune==0.8.4
nanotime==0.5.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
notebook==6.2.0
nonconformist==2.1.0
numba==0.52.0
numpy==1.20.1
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.2
pandas-profiling==2.11.0
pandocfilters==1.4.3
parso==0.8.1
pathlib2==2.3.5
pathspec==0.8.1
pexpect==4.8.0
phik==0.11.0
phonenumbers==8.12.18
pickleshare==0.7.5
Pillow==8.2.0
plotly==4.14.3
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.15.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.2
Pygments==2.8.0
pygtrie==2.3.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-benedict==0.23.2
python-dateutil==2.8.1
python-fsutil==0.4.0
python-slugify==4.0.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rich==9.12.0
rsa==4.7.2
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.1
Send2Trash==1.5.0
shortuuid==1.0.1
shtab==1.3.5
six==1.15.0
sklearn==0.0
smmap==3.0.5
tabulate==0.8.9
tangled-up-in-unicode==0.0.6
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.4
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.58.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.5
visions==0.6.0
voluptuous==0.12.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xmltodict==0.12.0
zc.lockfile==2.0
